Better warning, ChangeLog generation systems and docs fixes.
* Makefile.am:
* build/dist-changelog.am: Added better ChangeLog generation
system. Instead of depending on whole mm-common, a part of it was
taken and modified a bit.
* build/mm-standalone-warnings.m4:
* configure.ac: Use better warning system - also borrowed from
mm-common and modified a bit. In configure.ac - fixed .pc file
generation in AC_CONFIG_FILES_OUTPUT, added docs.sgml to it,
changed package name to libggraph.
* docs/reference/Makefile.am: sections file is again rebuilt, set
some ignores and so on.
* docs/reference/libggraph-docs.sgml.in: This file is now processed
by configure just to insert library and API versions.
* docs/reference/version.xml.in: Deleted.
* examples/Makefile.am: Removed EXTRA_DIST.
* examples/examples.am: Renamed from Makefile.am_fragment. Better
name and automagically distributed. Use better warning system.
* examples/simple_graph/Makefile.am:
* examples/simple_graph_construct/Makefile.am: Include examples.am
now.
* gsgraph/Makefile.am: Use better warning system.
* gsgraph/gsgraph*: Fix docs.